+++ docs/release-notes.md
@@ -9,6 +9,9 @@ Changes:
 * add two more known issues concerning the modal
 * add more robust argument and return value handling for lua parse() function
 * add issue template
+* rename namespace for components from `\BootstrapComponents\Component` to 
+    `\BootstrapComponents\Component` to comply with naming conventions
+* introduce class `ParserFirstCallInit` that handles the hook with the same name
 
 Fixes:
 * illegal call to User->loadFromSession() triggered by Extension:CodeMirror